handle score update during page save to catch screen add/remove
