+++ assets/src/apps/authoring/store/activities/actions/saveActivity.ts
@@ -41,37 +41,22 @@ export const saveActivity = createAsyncThunk(
         changeData,
         false,
       );
-      /* console.log('EDIT SAVE RESULTS', { editResults }); */
+
+      // grab the activity before it's updated for the score check
+      const oldActivityData = selectActivityById(rootState, activity.resourceId as number);
+
+      // update the activitiy before saving the page so that the score is correct
+      await dispatch(upsertActivity({ activity }));
+
       const currentPage = selectCurrentPage(rootState);
       if (!currentPage.custom.scoreFixed) {
         // need to check if this update to an activity affects the total score
-        const oldActivityData = selectActivityById(rootState, activity.resourceId as number);
         if (activity.content?.custom.maxScore !== oldActivityData?.content?.custom.maxScore) {
-          const allActivities = selectAllActivities(rootState);
-          const sequence = selectSequence(rootState);
-          const totalScore = sequence.reduce((acc, sequenceItem) => {
-            if (sequenceItem.custom.isLayer || sequenceItem.custom.isBank) {
-              return acc;
-            }
-            if (sequenceItem.resourceId === activity.resourceId) {
-              return acc + activity.content?.custom.maxScore;
-            }
-            const currActivity = allActivities.find((a) => a.id === sequenceItem.resourceId);
-            if (!currActivity) {
-              return acc;
-            }
-            return acc + (currActivity?.content?.custom?.maxScore || 0);
-          }, 0);
-          if (currentPage.custom.totalScore !== totalScore) {
-            console.log('updating total score', { totalScore, currentPage, activity });
-            await dispatch(savePage({ custom: { ...currentPage.custom, totalScore } }));
-            await dispatch(updatePage({ custom: { ...currentPage.custom, totalScore } }));
-          }
+          await dispatch(savePage());
         }
       }
     }
 
-    await dispatch(upsertActivity({ activity }));
     return;
   },
 );
