+++ assets/src/apps/authoring/store/page/actions/savePage.ts
@@ -1,11 +1,20 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
+import { selectAllActivities } from 'apps/delivery/store/features/activities/slice';
+import { selectSequence } from 'apps/delivery/store/features/groups/selectors/deck';
 import { ResourceContent } from 'data/content/resource';
 import { edit, Edited, ResourceUpdate } from 'data/persistence/resource';
 import { clone } from 'utils/common';
 import { selectAll as selectAllGroups } from '../../../../delivery/store/features/groups/slice';
 import { selectProjectSlug, selectReadOnly } from '../../app/slice';
 import { RootState } from '../../rootReducer';
-import { PageSlice, PageState, selectRevisionSlug, selectState, setRevisionSlug } from '../slice';
+import {
+  PageSlice,
+  PageState,
+  selectRevisionSlug,
+  selectState,
+  setRevisionSlug,
+  updatePage,
+} from '../slice';
 
 export const savePage = createAsyncThunk(
   `${PageSlice}/savePage`,
@@ -48,6 +57,26 @@ export const savePage = createAsyncThunk(
     };
     removeResourceId(updatedModel);
 
+    // need to update totalScore
+    const customUpdate = clone(payload.custom || currentPage.custom);
+    if (!customUpdate.scoreFixed) {
+      // need to calculate totalScore by all activities
+      const sequence = selectSequence(getState() as any);
+      const allActivities = selectAllActivities(getState() as any);
+      const totalScore = sequence.reduce((acc, sequenceItem) => {
+        if (sequenceItem.custom.isLayer || sequenceItem.custom.isBank) {
+          return acc;
+        }
+        const currActivity = allActivities.find((a) => a.id === sequenceItem.resourceId);
+        if (!currActivity) {
+          return acc;
+        }
+        return acc + (currActivity?.content?.custom?.maxScore || 0);
+      }, 0);
+      customUpdate.totalScore = totalScore;
+      dispatch(updatePage({ custom: customUpdate }));
+    }
+
     const update: ResourceUpdate = {
       title: payload.title || currentPage.title,
       objectives: payload.objectives || currentPage.objectives,
@@ -56,7 +85,7 @@ export const savePage = createAsyncThunk(
         advancedAuthoring,
         advancedDelivery,
         displayApplicationChrome,
-        custom: payload.custom || currentPage.custom,
+        custom: customUpdate,
         customCss: payload.customCss || currentPage.customCss,
         customScript: payload.customScript || currentPage.customScript,
         additionalStylesheets: payload.additionalStylesheets || currentPage.additionalStylesheets,
